@@ -1,5 +1,5 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
@@ -9,13 +9,15 @@ import store from './store'
 
 import { BrowserRouter as Router } from 'react-router-dom'
 
-ReactDOM.render(
+const container = document.getElementById('root')
+const root = createRoot(container)
+
+root.render(
   <Provider store={store}>
     <Router>
       <App />
     </Router>
-  </Provider>,
-  document.getElementById('root')
+  </Provider>
 )
 
 // If you want to start measuring performance in your app, pass a function
